+++ src/main/web/WEB-INF/web.xml
@@ -9,7 +9,7 @@
         <servlet-class>org.glassfish.jersey.servlet.ServletContainer</servlet-class>
         <init-param>
             <param-name>jersey.config.server.provider.packages</param-name>
-            <param-value>app.service</param-value>
+            <param-value>app.resources</param-value>
         </init-param>
         <load-on-startup>1</load-on-startup>
     </servlet>
